+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 ## [Unreleased]
 
 ### Fixed
+- fix config-dump systemtest [PR #736]
 - fix systemtests daemon control scripts [PR #762]
 - fix invalid file descriptor issue in the libcloud plugin [PR #702]
 - fix crash when loading both python-fd and python3-fd plugins [PR #730]
@@ -23,6 +24,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 - support for shorter date formats, where shorter dates are compensated with lowest value possible to make a full date [PR #707]
 
 ### Changed
+- bstrncpy: workaround when used with overlapping strings [PR #736]
 - Disabled test "statefile" for big endian, use temporary state files for all other architectures [PR #757]
 - Fixed broken link in https://docs.bareos.org/IntroductionAndTutorial/WhatIsBareos.html documentation page
 - Package **bareos-database-postgresql**: add recommendation for package **dbconfig-pgsql**.